requirements.txt
# YAML support
PyYAML>=6.0.1

# FastAPI + ASGI server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Settings management
pydantic-settings>=2.2.0

# AI / keyword extraction
groq>=0.9.0
yake>=0.4.8

# Database (async SQLAlchemy + PostgreSQL)
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.9

# Payments
razorpay>=1.4.0

# Testing
pytest>=8.1.0
pytest-asyncio>=0.23.0
httpx>=0.27.0         # async test client for FastAPI
